link to codesandbox demo of useContext example
To understand the example code in more detail, I implemented it in codesandbox. Here is a github repo if you want to fork it: https://github.com/100ideas/mobx-react-usecontext-codesandbox-demo

I was unable to get the original `createStore()` function to work as an observable, so I changed it slightly to an `observable(pojo)`. You should review it before accepting this PR to make sure what I changed is how you intend this pattern to be implemented.

```
 createStore() {
  // storing variable in closure means you can access it in methods
  // otherwise you would have to use `this`
  const store = {
//...
```